Move trigger logic into its own file.
